update 28/11/2021
update de manha

+++ trabalho/DAO/ClienteDAO.py
@@ -7,8 +7,8 @@ class ClienteDb(BancoDb):
     def __init__(self):
         pass
     
-    def inserirCliente(self, cliente):#PEGANDO INPUT DO FORM
-        if (cliente.getNome() == "" or cliente.getEndereco() == "" or cliente.getTelefone() == "" or cliente.getCpf() == ""):#nao aceita campo nulo
+    def inserirCliente(self, cliente):
+        if (cliente.getNome() == "" or cliente.getEndereco() == "" or cliente.getTelefone() == "" or cliente.getCpf() == ""):
             raise "All fields must be entered"
         else:
             dados = (cliente.getNome(), cliente.getEndereco(), cliente.getTelefone(), cliente.getCpf())
@@ -19,8 +19,8 @@ class ClienteDb(BancoDb):
             db.commit()
             db.close()    
             
-    def alterarCliente(self, cliente):#PEGANDO INPUT DO FORM
-        if (cliente.getNome() == "" or cliente.getEndereco() == "" or cliente.getTelefone() == ""):#nao aceita campo nulo
+    def alterarCliente(self, cliente):
+        if (cliente.getNome() == "" or cliente.getEndereco() == "" or cliente.getTelefone() == ""):
             raise "All fields must be entered"
         if self.verificaExistencia(cliente.getCpf()):
             dados = (cliente.getNome(), cliente.getEndereco(), cliente.getTelefone(), cliente.getCpf())
@@ -34,7 +34,7 @@ class ClienteDb(BancoDb):
         else:
             print("Não existe nenhum cliente com o CPF informado")
             
-    def consultarCliente(self, cpf): #CONSULTA PEGANDO INPUT DO FORM
+    def consultarCliente(self, cpf):
         db = MySQLdb.connect(self.banco_host, self.banco_username, self.banco_password, self.banco_nome)
         cursor = db.cursor()
         sql = "SELECT * FROM cliente_tbl WHERE cliente_cpf = %s;"%(cpf)
@@ -46,19 +46,36 @@ class ClienteDb(BancoDb):
         result_format = f"""NOME: {result[0]}   ENDERECO: {result[1]}   TELEFONE: {result[2]}   CPF: {result[3]}"""
         return result_format
             
-        
-    def deletarCliente(self, cpf):#FUNCIONANDO
+    def deletarCliente(self, cpf):
         if self.verificaExistencia(cpf):
             db = MySQLdb.connect(self.banco_host, self.banco_username, self.banco_password, self.banco_nome)
             cursor = db.cursor()
-            sql = """DELETE FROM cliente_tbl WHERE cliente_cpf = %s;"""%(cpf)
+            sql = f"select cliente_tbl.cliente_cpf from cliente_tbl inner join compra_tbl ON cliente_tbl.cliente_cpf = compra_tbl.cliente_cpf where cliente_tbl.cliente_cpf = {cpf}"
             cursor.execute(sql)
-            db.commit()
+            result = cursor.fetchall()
+            if len(result) == 0:
+                sql = """DELETE FROM cliente_tbl WHERE cliente_cpf = %s;"""%(cpf)
+                cursor.execute(sql)
+                db.commit()
+            else:
+                print("O cliente tem compras relacionadas a ele. Impossivel remover do sistema")
         else:
             print("Nao existe nenhum cliente com o CPF informado")
         db.close()
+     
         
-    def listarClientes(self): #FUNCIONANDO, PRINTA CADA UM LINHA POR LINHA NUMA LISTBOX
+#   def deletarCliente(self, cpf):
+#        if self.verificaExistencia(cpf):
+#           db = MySQLdb.connect(self.banco_host, self.banco_username, self.banco_password, self.banco_nome)
+#           cursor = db.cursor()
+#           sql = """DELETE FROM cliente_tbl WHERE cliente_cpf = %s;"""%(cpf)
+#           cursor.execute(sql)
+#           db.commit()
+#       else:
+#           print("Nao existe nenhum cliente com o CPF informado")
+#       db.close()
+       
+    def listarClientes(self):
         db = MySQLdb.connect(self.banco_host, self.banco_username, self.banco_password, self.banco_nome)
         cursor = db.cursor()
         sql = "SELECT * FROM cliente_tbl;"
